end-1-term-projects/phc/inspiring-phrase-generator-prototype/requirements.txt
beautifulsoup4==4.6.0
requests==2.18.4
pyserial==3.4